Data Module: Add default registry and createRegistry function
